- added error handling on frontend - added feedback when there is no topmovies - changed structure with better SOC
