+++ src/app.js
@@ -87,6 +87,9 @@ export default function initApp(api) {
   })
 
   app.use((request, response) => {
+    if (request.originalUrl.startsWith('/api/')) {
+      return response.status(404).json({ error: 'Endpoint hittades inte', status: 404 })
+    }
     renderErrorPage(response, 404, 'Sidan kunde inte hittas', 'Sidan finns inte')
   })
