+++ templates/layouts/main.ejs
@@ -72,7 +72,7 @@
 
   <script src="/static/script/celebration-modal.js"></script>
   <script src="/static/script/UI-dateFilter.js"></script>
-  <script type="module" src="/static/Script/topRatedMovies.js"></script>
+  <script type="module" src="/static/Script/main.js"></script>
 </body>
 
 </html>
